requirements.txt
Django>=5.0,<6
python-dotenv>=1.0.0
